[Improvement]: Optimizer-container refactor into ams/server (#2077)
* module refactor

* fix pom problem

* fix checkstyle
