+++ ams/optimizer/standalone-optimizer/pom.xml
@@ -21,10 +21,10 @@
          xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <parent>
-        <artifactId>amoro-ams</artifactId>
+        <artifactId>amoro-optimizer</artifactId>
         <groupId>com.netease.amoro</groupId>
         <version>0.6.0-SNAPSHOT</version>
-        <relativePath>../../pom.xml</relativePath>
+        <relativePath>../pom.xml</relativePath>
     </parent>
     <modelVersion>4.0.0</modelVersion>
