+++ ams/pom.xml
@@ -33,10 +33,7 @@
 
     <modules>
         <module>api</module>
-        <module>optimizer-container</module>
-        <module>optimizer/flink-optimizer</module>
-        <module>optimizer/standalone-optimizer</module>
-        <module>optimizer/common</module>
+        <module>optimizer</module>
         <module>server</module>
         <module>dashboard</module>
     </modules>
