+++ docs/admin-guides/managing-optimizers.md
@@ -25,7 +25,7 @@ Local container is a way to start Optimizer by local process and supports multi-
 ```yaml
 containers:
   - name: localContainer
-    container-impl: com.netease.arctic.optimizer.container.LocalOptimizerContainer
+    container-impl: com.netease.arctic.server.manager.LocalOptimizerContainer
     properties:
       export.JAVA_HOME: "/opt/java"   # JDK environment
 ```
@@ -33,7 +33,7 @@ containers:
 ### Flink container
 Flink container is a way to start Optimizer through Flink jobs. With Flink, you can easily deploy Optimizer
 on yarn clusters or kubernetes clusters to support large-scale data scenarios. To use flink container, 
-you need to add a new container configuration. with container-impl as `com.netease.arctic.optimizer.container.FlinkOptimizerContainer`
+you need to add a new container configuration. with container-impl as `com.netease.arctic.server.manager.FlinkOptimizerContainer`
 
 FlinkOptimizerContainer support the following properties:
 
@@ -64,7 +64,7 @@ An example for yarn-per-job mode:
 ```yaml
 containers:
   - name: flinkContainer
-    container-impl: com.netease.arctic.optimizer.container.FlinkOptimizerContainer
+    container-impl: com.netease.arctic.server.manager.FlinkOptimizerContainer
     properties:
       flink-home: /opt/flink/                              #flink install home
       export.HADOOP_CONF_DIR: /etc/hadoop/conf/            #hadoop config dir
@@ -78,7 +78,7 @@ An example for kubernetes-application mode:
 ```yaml
 containers:
   - name: flinkContainer
-    container-impl: com.netease.arctic.optimizer.container.FlinkOptimizerContainer
+    container-impl: com.netease.arctic.server.manager.FlinkOptimizerContainer
     properties:
       flink-home: /opt/flink/                                                        #flink install home
       target: kubernetes-application                                                 #flink run as native kubernetes
